LayoutTests/editing/pasteboard/4922709.html

<style>
blockquote {
    border-left: 2px solid blue;
    margin: 0;
    padding-left: 10px;
    color: blue;
}
</style>
<div id="description">This tests copying/pasting less than a paragraph of quoted content.  It should appear quoted.</div>
<div id="edit" contenteditable="true">
<br>
<br>
<div>On Tuesday, Dave wrote:</div>
<div><br></div>
<blockquote id="blockquote" type="cite">Hello World</blockquote>
</div>

<script>
if (window.layoutTestController)
    window.layoutTestController.dumpAsText();
blockquote = document.getElementById("blockquote");
text = blockquote.firstChild;
sel = window.getSelection();
sel.setBaseAndExtent(text, 0, text, text.length);
document.execCommand("Copy");
edit = document.getElementById("edit");
sel.setPosition(edit, 0);
document.execCommand("Paste");
if (window.layoutTestController)
    document.body.innerText = document.getElementById("description").innerText + "\n\n" + document.getElementById("edit").innerHTML;
</script>
