LayoutTests/editing/pasteboard/4922709-expected.txt

This tests copying/pasting less than a paragraph of quoted content. It should appear quoted.


<blockquote id="blockquote" type="cite" style="border-left-width: 2px; border-left-style: solid; border-left-color: blue; margin-top: 0px; margin-right: 0px; margin-bottom: 0px; margin-left: 0px; padding-left: 10px; color: blue; ">Hello World</blockquote>
<br>
<div>On Tuesday, Dave wrote:</div>
<div><br></div>
<blockquote id="blockquote" type="cite">Hello World</blockquote>
